Create the text preview feature and using map to loop the ContentCard
I have added the text preview feature which will cut the text at certain length, and I created a dummy JSON data that can be used for the mapping. I will use the API after the API is complete and fetch the data based on that. I will create the pagination feature next on the /results page.