Fix bugs on the summary page, revert the loading to previous function
Now, when you click back button in the summary page after you go to the author's rewind, it will go back to the results page, if you click the author on the results page, else it will go to the homepage.

Also, reverted the loading progress to the previous version, putting 200 instead of 404 on the backend return response, on the /progress route.